+++ src/Exception/InvalidArgumentException.php
@@ -10,6 +10,11 @@
 
 namespace LitGroup\Sms\Exception;
 
+/**
+ * Exception thrown if an argument does not match with the expected value.
+ *
+ * @author Roman Shamritskiy <roman@litgroup.ru>
+ */
 class InvalidArgumentException extends \InvalidArgumentException
 {
     public function __construct($message)